+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Report App Memory Usage (#2027)
+- Include app permissions with event (#1984)
 
 ## 7.23.0